+++ packages/genesys/packages/react-components-core/src/config/config.types.ts
@@ -40,6 +40,7 @@ export interface BaseCoreConfig {
     projectName: string;
     themes: Themes;
     translate: TranslateFn;
+    locale: string;
     validators: Record<keyof ExtraValidators, IValidator>;
     components: ComponentsConfig;
 }