+++ packages/genesys/packages/react-hooks/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Versions
 
+## 2.0.5
+
+Dependencies update
+
 ## 1.2.5
 
 Dependencies update