+++ packages/genesys/packages/react-native-components/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Versions
 
+## 2.0.24
+
+[Image] fix image props
+
 ## 2.0.23
 
 [Image] fix image props